fix #169 split View and ViewWrapper
